+++ lib/module.ex
@@ -244,7 +244,7 @@ defmodule Module do
   def defined_functions(module) do
     assert_not_compiled!(:defined_functions, module)
     table = function_table_for(module)
-    lc { tuple, _, _, _, _, _, _, _ } in ETS.tab2list(table), do: tuple
+    lc { tuple, _, _, _, _, _, _, _ } inlist ETS.tab2list(table), do: tuple
   end
 
   @doc """
@@ -263,7 +263,7 @@ defmodule Module do
   def defined_functions(module, kind) do
     assert_not_compiled!(:defined_functions, module)
     table = function_table_for(module)
-    lc { tuple, stored_kind, _, _, _, _, _, _ } in ETS.tab2list(table) when stored_kind == kind, do: tuple
+    lc { tuple, stored_kind, _, _, _, _, _, _ } inlist ETS.tab2list(table), stored_kind == kind, do: tuple
   end
 
   @doc """
@@ -274,7 +274,7 @@ defmodule Module do
   def make_overridable(module, tuples) do
     assert_not_compiled!(:make_overridable, module)
     table = function_table_for(module)
-    lc tuple in tuples do
+    lc tuple inlist tuples do
       case ETS.lookup(table, tuple) do
         [clause] ->
           ETS.delete(table, tuple)