+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 * backwards incompatible changes
   * [Builtin] `__using__` callback triggered by `use` now receives just one argument. Caller information can be accessed in the `__CALLER__`;
   * [Module] Removed data functions in favor of unifying the attributes API;
+  * [Kernel] Comprehensions syntax changed to be more compatible with Erlang behavior;
 
 * deprecations
   * [File] `File.read_info` was deprecated in favor of `File.stat`;